INS-3617: fix races
