+++ scripts/insolard/keeper_template.yaml
@@ -7,6 +7,6 @@ keeper:
   queryurl: 'https://prometheus.k8s-dev.insolar.io/api/v1/query?query='
   queries:
     - 'go_memstats_heap_inuse_bytes{installation="dev-alpha"} <bool 6000000000'
-    - 'insolar_requests_opened{installation="dev-alpha"} - insolar_requests_closed{installation="dev-alpha"} <bool 1000'
-    - 'rate(insolar_filament_length{installation="dev-alpha"}[1m]) <bool 5000'
+    - 'insolar_requests_opened{installation="dev-alpha"} - insolar_requests_closed{installation="dev-alpha"} <bool 5000'
+    - 'rate(insolar_filament_length{installation="dev-alpha"}[1m]) <bool 300000'
   maxmetriclag: 2m
